+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Options `-v` and `--version` display version number.
+
+### Fixed
+- Option parsing was broken. Switched to use [dashdash](https://github.com/trentm/node-dashdash) and added tests for
+option parsing.
+
 
 ## [0.1.0]
 ### Added